- Fix a bug where Moji.type? was returning true for unknown types. - Extract classes to their own files in Moji namespace.
