+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --exclude-limit 1000`
-# on 2023-04-20 19:59:18 UTC using RuboCop version 1.50.0.
+# on 2023-04-20 21:06:39 UTC using RuboCop version 1.50.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -24,21 +24,27 @@ Metrics/MethodLength:
 # Offense count: 1
 # Configuration parameters: CountComments, CountAsOne.
 Metrics/ModuleLength:
-  Max: 173
+  Max: 174
 
 # Offense count: 1
 # Configuration parameters: AllowedMethods, AllowedPatterns.
 Metrics/PerceivedComplexity:
   Max: 9
 
-# Offense count: 5
+# Offense count: 6
+RSpec/MultipleExpectations:
+  Max: 5
+
+# Offense count: 4
 # Configuration parameters: AllowedConstants.
 Style/Documentation:
   Exclude:
     - 'spec/**/*'
     - 'test/**/*'
     - 'lib/moji.rb'
-    - 'lib/moji/flag_set_maker.rb'
+    - 'lib/moji/detail.rb'
+    - 'lib/moji/flag_set.rb'
+    - 'lib/moji/flags.rb'
 
 # Offense count: 4
 # This cop supports safe autocorrection (--autocorrect).
